Add Order Limit to Pickup Events (#214)
Add a field to Pickup Events designating the total number of orders
that can be assigned to a given pickup event. This allows to reduce
logistical constraints that might occur from having too many orders
bunch up at one event.

This PR includes the field changes as well as a test suite ensuring
that consistency is kept with the order limit, along with ensuring
that we have descriptive errors for whenever events have reached
their order limit or other similar situations.